start writing man documentation
obviously as is the documentation is unusable. But once type requirements start to stabilize,
it'll be good to write man documentation. It's much easier to quickly look up a manpage than dig through
a header file, plus, I don't really want to bombard the type requirement headers with rationale for
including/not including behaviors