starting singleList
I got tired of mentally trying to keep track of what would be ok for doubleList to take from singleList
and seeing what parts would be ok if the list is circular. So I guess we're just writing singleList first,
which is not really what I wanted (I just want to freaking write fheap). But anyways this is the only organized
way to do things...